create validate id inside token in route sponsor agent

+++ src/modules/Agents/UseCases/ToCancelSponsorAgent/ToCancelSponsorAgentUseCase.ts
@@ -7,9 +7,9 @@ class ToCancelSponsorAgentUseCase{
   constructor(sponsorAgentRepository:ISponsorAgentRepository) {
     this.sponsorAgentRepository = sponsorAgentRepository
   }
-  async execute({ id_agent, id_sponsor }): Promise<SponsorAgent> {
+  async execute({id_agent_token, id_agent, id_sponsor }): Promise<SponsorAgent> {
     if(!id_agent || !id_sponsor) throw new AppError("invalid parameters")
-
+    if(id_agent !== id_agent_token) throw new AppError("Token sen not to own agent authenticate ")
     const canceledSponsor = await this.sponsorAgentRepository.delete({ id_agent, id_sponsor })
     return canceledSponsor
   }
